+++ Common/src/main/java/fuzs/easyshulkerboxes/EasyShulkerBoxes.java
@@ -1,8 +1,10 @@
 package fuzs.easyshulkerboxes;
 
 import fuzs.easyshulkerboxes.config.ClientConfig;
+import fuzs.easyshulkerboxes.config.ServerConfig;
 import fuzs.easyshulkerboxes.init.ModRegistry;
 import fuzs.easyshulkerboxes.network.client.message.C2SEnderChestMenuMessage;
+import fuzs.easyshulkerboxes.network.client.message.C2SEnderChestSetSlotMessage;
 import fuzs.easyshulkerboxes.network.message.S2CEnderChestSetContentMessage;
 import fuzs.easyshulkerboxes.network.message.S2CEnderChestSetSlotMessage;
 import fuzs.puzzleslib.config.ConfigHolder;
@@ -20,7 +22,9 @@ public class EasyShulkerBoxes implements ModConstructor {
 
     public static final NetworkHandler NETWORK = CoreServices.FACTORIES.network(MOD_ID);
     @SuppressWarnings("Convert2MethodRef")
-    public static final ConfigHolder CONFIG = CoreServices.FACTORIES.clientConfig(ClientConfig.class, () -> new ClientConfig());
+    public static final ConfigHolder CONFIG = CoreServices.FACTORIES
+            .clientConfig(ClientConfig.class, () -> new ClientConfig())
+            .serverConfig(ServerConfig.class, () -> new ServerConfig());
 
     @Override
     public void onConstructMod() {
@@ -32,6 +36,7 @@ public class EasyShulkerBoxes implements ModConstructor {
     private static void registerMessages() {
         NETWORK.register(S2CEnderChestSetContentMessage.class, S2CEnderChestSetContentMessage::new, MessageDirection.TO_CLIENT);
         NETWORK.register(S2CEnderChestSetSlotMessage.class, S2CEnderChestSetSlotMessage::new, MessageDirection.TO_CLIENT);
+        NETWORK.register(C2SEnderChestSetSlotMessage.class, C2SEnderChestSetSlotMessage::new, MessageDirection.TO_SERVER);
         NETWORK.register(C2SEnderChestMenuMessage.class, C2SEnderChestMenuMessage::new, MessageDirection.TO_SERVER);
     }
 }