+++ Common/src/main/java/fuzs/easyshulkerboxes/handler/EnderChestMenuHandler.java
@@ -15,6 +15,8 @@ public class EnderChestMenuHandler {
 
     public void onLivingTick(LivingEntity entity) {
         if (entity instanceof ServerPlayer) {
+            // this should be ok to do every tick as only actual changes are sent
+            // vanilla also does this every tick for the current menu (inventory most of the time)
             ModRegistry.ENDER_CHEST_MENU_CAPABILITY.maybeGet(entity)
                     .map(EnderChestMenuCapability::getEnderChestMenu)
                     .ifPresent(AbstractContainerMenu::broadcastChanges);
@@ -23,6 +25,7 @@ public class EnderChestMenuHandler {
 
     public static Optional<AbstractContainerMenu> openEnderChestMenu(Player player) {
         return ModRegistry.ENDER_CHEST_MENU_CAPABILITY.maybeGet(player).map(capability -> {
+            // container id doesn't matter since we do the syncing ourselves where the id is never used
             ChestMenu menu = ChestMenu.threeRows(-100, new Inventory(player), player.getEnderChestInventory());
             capability.setEnderChestMenu(menu);
             return menu;